Refactor Helm auth
auth components moved to discover-auth
chart and used as a subchart. This enables the
same chart to be used with search-abstractor.

Some attempt to sanitize the various discover/values files
in relation to discover-auth.

discover-auth chart changes:
* Use generated README.md
* Document values.yaml
* Use preexisting imagePullSecrets
* Remove use of nodePort

Would be useful to extend some of this to the rest of the
discover helm chart, but out of scope for now.

Signed-off-by: daniel-lau-mf <36928012+daniel-lau-mf@users.noreply.github.com>